src/test/java/com/amnesty/panicbutton/ApplicationSettingsTest.java
package com.amnesty.panicbutton;

import android.app.Application;
import android.content.SharedPreferences;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.robolectric.Robolectric;
import org.robolectric.RobolectricTestRunner;
import org.robolectric.shadows.ShadowPreferenceManager;

import static com.amnesty.panicbutton.ApplicationSettings.completeFirstRun;
import static com.amnesty.panicbutton.ApplicationSettings.isFirstRun;
import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

@RunWith(RobolectricTestRunner.class)
public class ApplicationSettingsTest {

    private SharedPreferences sharedPreferences;
    private Application context;

    @Before
    public void setUp() {
        context = Robolectric.application;
        sharedPreferences = ShadowPreferenceManager.getDefaultSharedPreferences(context);
    }

    @Test
    public void shouldReturnTrueWhenFlagNotFound() {
        assertTrue(isFirstRun(context));
    }

    @Test
    public void shouldSaveFalseAsFlagOnFirstRunCompletion() {
        completeFirstRun(context);
        assertFalse(sharedPreferences.getBoolean("FIRST_RUN", true));
    }

    @Test
    public void shouldReturnFalseWhenFirstRunCompleted() {
        SharedPreferences.Editor editor = sharedPreferences.edit();
        editor.putBoolean("FIRST_RUN", false);
        editor.commit();
        assertFalse(isFirstRun(context));
    }
}
